+++ src/haddock/libs/libcns.py
@@ -24,23 +24,22 @@ def generate_default_header(
     path: Optional[FilePath] = None,
 ) -> tuple[str, str, str, str, str, str]:
     """Generate CNS default header."""
-    # TODO: Remove the `type: ignore` comments
     if path is not None:
-        axis = load_axis(**cns_paths.get_axis(path))  # type: ignore
+        axis = load_axis(**cns_paths.get_axis(path))
         link = load_link(Path(path, cns_paths.LINK_FILE))
         scatter = load_scatter(Path(path, cns_paths.SCATTER_LIB))
-        tensor = load_tensor(**cns_paths.get_tensors(path))  # type: ignore
+        tensor = load_tensor(**cns_paths.get_tensors(path))
         trans_vec = load_trans_vectors(
-            **cns_paths.get_translation_vectors(path)  # type: ignore
-        )  # noqa: E501
+            **cns_paths.get_translation_vectors(path)
+        )
         water_box = load_boxtyp20(cns_paths.get_water_box(path)["boxtyp20"])
 
     else:
-        axis = load_axis(**cns_paths.axis)  # type: ignore
+        axis = load_axis(**cns_paths.axis)
         link = load_link(cns_paths.link_file)
         scatter = load_scatter(cns_paths.scatter_lib)
-        tensor = load_tensor(**cns_paths.tensors)  # type: ignore
-        trans_vec = load_trans_vectors(**cns_paths.translation_vectors)  # type: ignore
+        tensor = load_tensor(**cns_paths.tensors)
+        trans_vec = load_trans_vectors(**cns_paths.translation_vectors)
         water_box = load_boxtyp20(cns_paths.water_box["boxtyp20"])
 
     return (
@@ -93,7 +92,7 @@ def find_desired_linkfiles(
     # Point to corresponding file
     linkfiles["prot_link_infile"] = cns_paths.PROTEIN_LINK_FILES[prot_link_key]
 
-    # Logic to find linkfile for dna
+    # Logic to find linkfile for DNA/RNA
     nucl_link_key = "5'Phosphate" if phosphate_5 else "5'OH"
     # Point to corresponding file
     linkfiles["nucl_link_infile"] = cns_paths.NUCL_LINK_FILES[nucl_link_key]
@@ -392,21 +391,22 @@ def prepare_cns_input(
                     "[topocg].")
                 else:
                     aa_psf_list.append(psf.rel_path.as_posix())
-            for tbl in input_element.cgtoaa_tbl :
-                if tbl is None:
+            for i, tbl in enumerate(input_element.cgtoaa_tbl):
+                if tbl is None and not input_element.shape[i]:
                     raise ValueError(f"Coarse-Crain to All-Atom restraint file not found "
                     "{input_element.rel_path}. Conversion to all-atom requires a restraint file "
                     "generated with [topocg].")
-                else:
+                elif not input_element.shape[i]:
                     cgtoaa_tbl_list.append(tbl.as_posix())
         else:
             pdb = input_element
+            shape = libpdb.check_mol_shape(pdb.rel_path)
             if pdb.aa_topology is None:
                 raise ValueError(f"All-Atom Topology not found {input_element.rel_path}."
                 "Conversion to all-atom requires a topology generated with [topoaa] and "
                 "[topocg].")
             aa_psf_list.append(pdb.aa_topology.rel_path.as_posix())
-            if pdb.cgtoaa_tbl is None:
+            if pdb.cgtoaa_tbl is None and not shape:
                 raise ValueError(f"Coarse-Crain to All-Atom restraint file not found for"
                 " entry: {input_element.rel_path}. Conversion to all-atom requires a restraint file "
                 "generated with [topocg].")
@@ -485,9 +485,11 @@ def prepare_expected_pdb(
         pdb.topology = [p.topology for p in model_obj]
         pdb.aa_topology = [p.aa_topology for p in model_obj]
         pdb.cgtoaa_tbl = [p.cgtoaa_tbl for p in model_obj]
+        pdb.shape = [p.shape for p in model_obj]
     else:
         pdb.topology = model_obj.topology
         pdb.seed = model_obj.seed
         pdb.aa_topology = model_obj.aa_topology
         pdb.cgtoaa_tbl = model_obj.cgtoaa_tbl
+        pdb.shape = model_obj.shape
     return pdb